+++ purchase_work_acceptance_disbursement/__manifest__.py
@@ -9,6 +9,7 @@
     'depends': ['purchase_work_acceptance_kmitl', 'purchase_order_disbursement'],
     "data": [
         "views/disbursement_request_views.xml",
+        "views/purchase_order_views.xml",
         "views/work_acceptance_views.xml"
     ],
     'installable': True,
